fix undefined behavior caused by aliased references
When dropping and reusing a "Node" we used to create a "&mut Node<T>", but other threads may create "& Node<T>" at the same time to read from the atomic reference count. The existence of such two references at the same time is undefined behavior. To avoid this we only ever use pointers to the node and only create references for the nodes fields.